pyproject.toml
[tool.poetry]
name = "cosy"
version = "2.0.0"
description = ""
authors = ["Finn Stoldt <f.stoldt@uni-luebeck.de>", "Jakob Behnke <j.behnke@uni-luebeck.de>"]
readme = "README.md" 

[tool.poetry.dependencies]
python = "3.10.13"
Flask = "^3.0.2"
PyAudio = "^0.2.14"
pymitter = "^0.5.0"
watchdog = "^4.0.0"
PyYAML = "^6.0.1"
psutil = "^5.9.8"
numpy = "^1.26.4"
requests = "2.31.0"
flask-cors = "^4.0.0"
pydub = "^0.25.1"
laboratorium-ai-pitch-loudness = {version = "^1.0.2", source = "uzl"}
laboratorium-ai-nlp = {version = "^1.0.1", source = "uzl"}
laboratorium-ai-forced-alignment = {version = "^1.1.3", source = "uzl"}
laboratorium-ai-asr = {version = "^1.0.2", source = "uzl"}

[tool.poetry.scripts]
main = "cosy.__main__:run"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[[tool.poetry.source]]
name = "uzl"
url = "https://maven.itm.uni-luebeck.de/repository/pypi/simple"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
